ntpd: create config_file directory
ntpd's configuration is stored in /var/etc, but the init script doesn't
create this directory. To be fair this folder is used by other packages
as well, for instance dnsmasq. Still, it's better not to rely on other
packages to create it.

Fixes: #10636

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>
(cherry picked from commit 18335dc4241a490ae9e0100355492da0c5012e4b)